add fixes

+++ src/main/java/com/mariusz/contacts/mapper/CustomerMapper.java
@@ -9,7 +9,7 @@ import java.sql.SQLException;
 public class CustomerMapper implements RowMapper<Customer> {
     public Customer mapRow(ResultSet rs, int rowNum) throws SQLException {
             Customer customer = new Customer();
-            customer.setId(rs.getInt("id"));
+            customer.setId(rs.getLong("id"));
             customer.setName(rs.getString("name"));
             customer.setSurname(rs.getString("surname"));
             customer.setAge(rs.getString("age"));
